migrations/requirements.txt
faker
mysql.connector